Start new connection for the #get_connection method
For the device and device_group providers, the `.get_connection` methods simply returned
`@connections[account]`, which would possibly result in `nil` being returned. This has
been updated to return `@connections[account]` or set the value to the result of
`start_connection "#{account}.logicmonitor.com"` and return it if the value is `nil`.